+++ unittest/hypergraph/hypergraph_test.cpp
@@ -1,0 +1,45 @@
+#include "pmondriaan.hpp"
+
+#include "gtest/gtest.h"
+
+#include <bulk/bulk.hpp>
+#ifdef BACKEND_MPI
+#include <bulk/backends/mpi/mpi.hpp>
+using environment = bulk::mpi::environment;
+#else
+#include <bulk/backends/thread/thread.hpp>
+using environment = bulk::thread::environment;
+#endif
+
+namespace pmondriaan {
+namespace {
+
+std::string mtx_three_nonzeros = R"(%%MatrixMarket matrix coordinate real general
+3 3 4
+1 1 1.0
+2 1 1.0
+2 2 1.0
+1 3 1.0
+)";
+
+TEST(Cutsize, CutsizeCutnet) {
+    environment env;
+    env.spawn(2, [](bulk::world& world) {
+        std::stringstream mtx_ss(mtx_three_nonzeros);
+        auto hypergraph = read_hypergraph_istream(mtx_ss, world, "degree");
+        auto H = hypergraph.value();
+        if (world.rank() == 0) {
+            H(0).set_part(0);
+            H(1).set_part(1);
+        }
+        if (world.rank() == 1) {
+            H(0).set_part(1);
+        }
+        auto cut = pmondriaan::cutsize(world, H, pmondriaan::m::cut_net);
+        ASSERT_EQ(cut, 2);
+    });
+}
+
+} // namespace
+} // namespace pmondriaan
+
